# backend/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisHeader": {
        "properties": {
          "analyzed_jobs": {
            "title": "Analyzed Jobs",
            "type": "integer"
          },
          "low_confidence": {
            "title": "Low Confidence",
            "type": "boolean"
          },
          "min_sample_size": {
            "title": "Min Sample Size",
            "type": "integer"
          },
          "salary_disclosed": {
            "title": "Salary Disclosed",
            "type": "integer"
          }
        },
        "required": [
          "analyzed_jobs",
          "salary_disclosed",
          "low_confidence",
          "min_sample_size"
        ],
        "title": "AnalysisHeader",
        "type": "object"
      },
      "AnalysisResponse": {
        "properties": {
          "arrangements": {
            "items": {
              "$ref": "#/components/schemas/DistributionBucket"
            },
            "title": "Arrangements",
            "type": "array"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryStats"
            },
            "title": "Categories",
            "type": "array"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "experience_levels": {
            "items": {
              "$ref": "#/components/schemas/DistributionBucket"
            },
            "title": "Experience Levels",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/JobFilters"
          },
          "header": {
            "$ref": "#/components/schemas/AnalysisHeader"
          },
          "salary": {
            "$ref": "#/components/schemas/SalarySummary"
          },
          "top_companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyStat"
            },
            "title": "Top Companies",
            "type": "array"
          }
        },
        "required": [
          "filters",
          "header",
          "categories",
          "arrangements",
          "experience_levels",
          "salary",
          "top_companies",
          "computed_at",
          "data_version"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "CategoryStats": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "distinct_technologies": {
            "title": "Distinct Technologies",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "technologies": {
            "items": {
              "$ref": "#/components/schemas/TechnologyStat"
            },
            "title": "Technologies",
            "type": "array"
          }
        },
        "required": [
          "category",
          "label",
          "technologies",
          "distinct_technologies"
        ],
        "title": "CategoryStats",
        "type": "object"
      },
      "CoOccurrenceStat": {
        "description": "A technology that appears alongside this skill more often than its\nbase rate in the same scope predicts (lift >= the floor).",
        "properties": {
          "baseline_share": {
            "title": "Baseline Share",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "lift": {
            "title": "Lift",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share_given_tech": {
            "title": "Share Given Tech",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "category",
          "count",
          "share_given_tech",
          "baseline_share",
          "lift"
        ],
        "title": "CoOccurrenceStat",
        "type": "object"
      },
      "CompanyStat": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "count"
        ],
        "title": "CompanyStat",
        "type": "object"
      },
      "DistributionBucket": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "share": {
            "title": "Share",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "count",
          "share"
        ],
        "title": "DistributionBucket",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobFilters": {
        "additionalProperties": false,
        "description": "Everything is optional; the empty filter means the national market.",
        "properties": {
          "arrangement": {
            "anyOf": [
              {
                "enum": [
                  "onsite",
                  "hybrid",
                  "remote"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "experience_level": {
            "anyOf": [
              {
                "enum": [
                  "entry",
                  "mid",
                  "senior",
                  "staff_plus"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "radius_miles": {
            "default": 50,
            "title": "Radius Miles",
            "type": "integer"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "technologies": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Technologies",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "JobFilters",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Login does not re-validate password length: complexity rules change\nover time and must never lock an existing account out of signing in.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "SalarySummary": {
        "properties": {
          "disclosed_count": {
            "title": "Disclosed Count",
            "type": "integer"
          },
          "median": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median"
          },
          "p25": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P25"
          },
          "p75": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P75"
          }
        },
        "required": [
          "disclosed_count",
          "p25",
          "median",
          "p75"
        ],
        "title": "SalarySummary",
        "type": "object"
      },
...
      "SignupRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "SkillDetailResponse": {
        "properties": {
          "arrangements": {
            "items": {
              "$ref": "#/components/schemas/DistributionBucket"
            },
            "title": "Arrangements",
            "type": "array"
          },
          "co_occurring": {
            "items": {
              "$ref": "#/components/schemas/CoOccurrenceStat"
            },
            "title": "Co Occurring",
            "type": "array"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "experience_levels": {
            "items": {
              "$ref": "#/components/schemas/DistributionBucket"
            },
            "title": "Experience Levels",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/JobFilters"
          },
          "header": {
            "$ref": "#/components/schemas/SkillHeader"
          },
          "requirement_levels": {
            "items": {
              "$ref": "#/components/schemas/DistributionBucket"
            },
            "title": "Requirement Levels",
            "type": "array"
          },
          "salary": {
            "$ref": "#/components/schemas/SkillSalary"
          },
          "trend": {
            "$ref": "#/components/schemas/SkillTrend"
          }
        },
        "required": [
          "filters",
          "header",
          "requirement_levels",
          "arrangements",
          "experience_levels",
          "co_occurring",
          "salary",
          "trend",
          "computed_at",
          "data_version"
        ],
        "title": "SkillDetailResponse",
        "type": "object"
      },
      "SkillHeader": {
        "properties": {
          "analyzed_jobs": {
            "title": "Analyzed Jobs",
            "type": "integer"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "jobs_with_tech": {
            "title": "Jobs With Tech",
            "type": "integer"
          },
          "low_confidence": {
            "title": "Low Confidence",
            "type": "boolean"
          },
          "min_sample_size": {
            "title": "Min Sample Size",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share": {
            "title": "Share",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "category",
          "analyzed_jobs",
          "jobs_with_tech",
          "share",
          "low_confidence",
          "min_sample_size"
        ],
        "title": "SkillHeader",
        "type": "object"
      },
...
      "SkillTrend": {
        "properties": {
          "days_observed": {
            "title": "Days Observed",
            "type": "integer"
          },
          "geo_slug": {
            "title": "Geo Slug",
            "type": "string"
          },
          "min_days": {
            "title": "Min Days",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "status": {
            "enum": [
              "ok",
              "collecting_history"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "days_observed",
          "min_days",
          "geo_slug",
          "points"
        ],
        "title": "SkillTrend",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Xcelsior API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis": {
      "get": {
        "operationId": "analysis_api_v1_analysis_get",
        "parameters": [
          {
            "description": "Free-text query (websearch syntax)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text query (websearch syntax)",
              "title": "Q"
            }
          },
          {
            "description": "Canonical title family slug",
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical title family slug",
              "title": "Title"
            }
          },
          {
            "description": "City slug, e.g. san-francisco-ca",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City slug, e.g. san-francisco-ca",
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "radius_miles",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Radius Miles",
              "type": "integer"
            }
          },
          {
            "description": "Technology slugs (AND)",
            "in": "query",
            "name": "tech",
            "required": false,
            "schema": {
              "default": [],
              "description": "Technology slugs (AND)",
              "items": {
                "type": "string"
              },
              "title": "Tech",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "arrangement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "onsite",
                    "hybrid",
                    "remote"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arrangement"
            }
          },
          {
            "in": "query",
            "name": "experience_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "entry",
                    "mid",
                    "senior",
                    "staff_plus"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experience Level"
            }
          },
          {
            "description": "Annual USD floor",
            "in": "query",
            "name": "salary_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Annual USD floor",
              "title": "Salary Min"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Exchange email + password for a session cookie. A wrong email or\npassword is an indistinguishable 401 (no account-enumeration signal).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/signup": {
      "post": {
        "description": "Create an account and sign the new user in (a session cookie is set),\nso the client can move straight into resume onboarding.",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/skills/{slug}": {
      "get": {
        "operationId": "skill_detail_api_v1_skills__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Free-text query (websearch syntax)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text query (websearch syntax)",
              "title": "Q"
            }
          },
          {
            "description": "Canonical title family slug",
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical title family slug",
              "title": "Title"
            }
          },
          {
            "description": "City slug, e.g. san-francisco-ca",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City slug, e.g. san-francisco-ca",
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "radius_miles",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Radius Miles",
              "type": "integer"
            }
          },
          {
            "description": "Technology slugs (AND)",
            "in": "query",
            "name": "tech",
            "required": false,
            "schema": {
              "default": [],
              "description": "Technology slugs (AND)",
              "items": {
                "type": "string"
              },
              "title": "Tech",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "arrangement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "onsite",
                    "hybrid",
                    "remote"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arrangement"
            }
          },
          {
            "in": "query",
            "name": "experience_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "entry",
                    "mid",
                    "senior",
                    "staff_plus"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experience Level"
            }
          },
          {
            "description": "Annual USD floor",
            "in": "query",
            "name": "salary_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Annual USD floor",
              "title": "Salary Min"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skill Detail",
        "tags": [
          "analytics"
        ]
      }
    }
  }
}